# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AnalystGuru — AI System Analyst Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all documents
      parameters:
        - name: doc_type
          in: query
          schema:
            type: ["string", "null"]
        - name: project_name
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a new document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── AI Reviewer ─────────────────────────────────────────────────────────────
...
  /kb/documents:
    get:
      operationId: listKbDocuments
      tags: [kb]
      summary: List KB documents
      responses:
        "200":
          description: KB documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: addKbDocument
      tags: [kb]
      summary: Add document to knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: KB document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
  /documents/{id}/generate-diagrams:
    post:
      operationId: generateDiagrams
      tags: [diagrams]
      summary: Generate all diagrams for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated diagram set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagramSet"
  /diagrams/{id}:
    get:
      operationId: getDiagram
      tags: [diagrams]
      summary: Get diagram by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Diagram artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagramArtifact"
  /diagrams/generate-c4:
    post:
      operationId: generateC4Diagrams
      tags: [diagrams]
      summary: Generate C4 diagrams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagramInput"
      responses:
        "200":
          description: C4 diagrams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagramSet"
  /diagrams/generate-uml:
    post:
      operationId: generateUmlDiagrams
      tags: [diagrams]
      summary: Generate UML diagrams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagramInput"
      responses:
        "200":
          description: UML diagrams
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagramSet"
  /diagrams/generate-erd:
    post:
      operationId: generateErdDiagram
      tags: [diagrams]
      summary: Generate ERD diagram
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagramInput"
      responses:
        "200":
          description: ERD diagram
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagramSet"

  # ─── ADR Records ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Document:
      type: object
      required: [id, created_at, title, text, doc_type]
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        title:
          type: string
        text:
          type: string
        doc_type:
          type: string
          enum: [tz, brd, user_story, srs, kb_article, urs, adr, api_spec]
        project_name:
          type: ["string", "null"]
        is_kb:
          type: boolean
          default: false
    DocumentInput:
      type: object
      required: [title, text]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        text:
          type: string
          minLength: 10
          maxLength: 30000
        doc_type:
          type: string
          default: tz
        project_name:
          type: ["string", "null"]
        is_kb:
          type: boolean
          default: false

    # ─── Review ───────────────────────────────────────────────────────────────
...
    DiagramArtifact:
      type: object
      required: [id, created_at, document_id, diagram_type, notation, source_code]
      properties:
        id:
          type: string
        created_at:
          type: string
          format: date-time
        document_id:
          type: string
        diagram_type:
          type: string
          enum: [c4_context, c4_container, c4_component, sequence, class, erd, use_case, activity, state, deployment, flowchart]
        notation:
          type: string
          enum: [plantuml, mermaid]
        source_code:
          type: string
        needs_review:
          type: boolean
    DiagramSet:
      type: object
      required: [document_id, diagrams, needs_review]
      properties:
        document_id:
          type: string
        diagrams:
          type: array
          items:
            $ref: "#/components/schemas/DiagramArtifact"
        c4_context:
          type: ["string", "null"]
        c4_container:
          type: ["string", "null"]
        c4_component:
          type: ["string", "null"]
        use_case:
          type: ["string", "null"]
        sequence:
          type: ["string", "null"]
        class_diagram:
          type: ["string", "null"]
        erd:
          type: ["string", "null"]
        mermaid_flowchart:
          type: ["string", "null"]
        confidence:
          type: string
          enum: [high, medium, low]
        needs_review:
          type: boolean
    DiagramInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 10
        diagram_types:
          type: array
          items:
            type: string

    # ─── Audit ────────────────────────────────────────────────────────────────
...