# lib/api-zod/src/generated/types/diagramArtifact.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnalystGuru — AI System Analyst Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { DiagramArtifactDiagramType } from './diagramArtifactDiagramType';
import type { DiagramArtifactNotation } from './diagramArtifactNotation';

export interface DiagramArtifact {
  id: string;
  created_at: Date;
  document_id: string;
  diagram_type: DiagramArtifactDiagramType;
  notation: DiagramArtifactNotation;
  source_code: string;
  needs_review?: boolean;
}